requirements.txt
# Web app
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
jinja2>=3.1.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Config
python-dotenv>=1.0.0
